trading_bot/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-binance==1.0.19
pydantic==2.7.4
